moved pids into pumps and mixer, hysteresis of boiler charging

+++ cTemp.h
@@ -99,7 +99,7 @@ class cTempSensor
 	
 	unsigned long StartTime;
 	  
-	float TempFilt;
+	double TempFilt;
 	const float* Offset;
 	
 	const int* pinMultiplexInput;
